+++ init.rb
@@ -7,7 +7,7 @@ Dir[File.dirname(__FILE__) + '/locale/**/*.yml'].each do |file|
   I18n.load_path << file
 end
 
-Tog::Plugins.settings :tog_wall,  :pagination_size => 50
+Tog::Plugins.settings :tog_wall,  'wall.pagination_size' => 10
 Tog::Plugins.settings :tog_wall,  'search.skip_indices' => false
 
 Tog::Plugins.helpers WallsHelper