+++ CHANGELOG.markdown
@@ -1,6 +1,8 @@
 EDGE
 ----
 * Added like comments functionality
+* Added flow pagination with will_paginate
+* Graffities and replys through ajax
 
 0.1.0
 ----